Updates external links and social network checks
Changes the 'Return to WealthWise' link to point to a new URL.

Improves the rendering logic for social network links by adding
a check for the existence and length of the SOCIALNETWORKS array
before attempting to map over it. This prevents potential errors if
the array is undefined or empty.

Removes the `transition:persist` attribute from a div.